remove standalone test
